apps/backend/api/v1/routes/models/form_requests: type of 'cargoWeight' param from 'int' to 'float'

+++ apps/backend/api/v1/routes/models/form_requests.py
@@ -8,6 +8,6 @@ class CalculateFormRequest(BaseModel):
     dispatchDate: datetime.date
     departureId: dict[str, Any]
     destinationId: dict[str, Any]
-    cargoWeight: int
+    cargoWeight: float
     containerType: int
     onlyInSelectedDateRange: bool = False
